split rules into folder
